+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * `LoadImage` trait
 * `DrawImage` trait
+* `HasSize` trait
 * Implementation of those traits for the types in [`playdate-sys`](https://docs.rs/playdate-sys/0.2) behind the feature
 flag `playdate-sys-v02` (enabled by default).